Refactor: main.js -> car.js 파일 이름 변경
